add document save to service document service
